requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.0
requests==2.31.0
scikit-learn==1.3.0
numpy==1.24.0
pandas==2.0.0
Werkzeug==3.0.0
PyJWT==2.8.0
cryptography==41.0.0
pymongo==4.5.0
SQLAlchemy==2.0.0
Gunicorn==21.2.0
